requirements.txt
Flask>=2.2
numpy
pandas
scikit-learn
Pillow
matplotlib
seaborn
joblib
python-dotenv
pyyaml
streamlit
